% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addParameter.R
\name{addParameter}
\alias{addParameter}
\title{addParameter}
\usage{
addParameter(param.name, value, param.dataframe)
}
\arguments{
\item{param.name}{name of the parameter}

\item{value}{value of the parameter}

\item{param.dataframe}{dataframe into which the parameter is to be inserted}
}
\value{
parameter dataframe with the inserted variable
}
\description{
This function inserts a new parameter value into an existing IPM parameter dataframe
using its name to locate the appropriate location for integration with the IPM.
}
\examples{

}
