% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_force.R
\docType{data}
\name{center_force}
\alias{center_force}
\title{Center all particles around the origin without affecting velocity}
\description{
This force repositions the particles at each generation so they are centered
around (0,0). It does not affect the velocity of the particles and are thus
mainly a guard against the whole body of particles drifting off.
}
\section{Training parameters}{

There are no parameters for this force.
}

\seealso{
Other forces: \code{\link{collision_force}},
  \code{\link{link_force}}, \code{\link{manybody_force}},
  \code{\link{mean_force}}, \code{\link{random_force}},
  \code{\link{reset_force}}, \code{\link{trap_force}},
  \code{\link{x_force}}, \code{\link{y_force}}
}
\keyword{datasets}
