% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clearParamFrame.R
\name{clearParamFrame}
\alias{clearParamFrame}
\title{Creates a new dataframe for IPM parameters}
\usage{
clearParamFrame(default = NA)
}
\arguments{
\item{default}{defaults to a 10-parameter dataframe - will be updated for generalisation later}
}
\value{
Empty dataframe of parameters to be populated for each run of the IPM
}
\description{
Create a new dataframe for IPM parameters
}
\examples{

}
