% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMrecruitsize.R
\name{conditionalSize}
\alias{conditionalSize}
\title{Size of Offspring}
\usage{
conditionalSize(x1, recruitMean, recruitSD)
}
\arguments{
\item{x1}{vector of individuals of state variable x (e.g., size) at time = t + 1 (new individuals at time t + 1)}

\item{recruitMean}{mean of the state variable (e.g., size) of new individuals recruited into the population at time t+1}

\item{recruitSD}{sd of the state variable (e.g., size) of new individuals recruited into the population at time t+1}
}
\value{
probability of establishment/recruitment
}
\description{
This function describes the conditional size distribution of new offspring
(i.e., offspring born at time time = t when observed at time = t+1)
}
