% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutLimit.R
\name{cutLimit}
\alias{cutLimit}
\title{cutLimit - Identify the minimum index of parameter values to eliminate}
\usage{
cutLimit(surface.x, decay.seq, threshold)
}
\arguments{
\item{decay.seq}{Decay sequence generated by paramDecay(surface.x, surface.y)}

\item{threshold}{Percentage threshold by which one wants the domain limited}
}
\value{
The identity of the minimum number of parametere state values that must be
eliminated based on performance to decrease the parameter domain by threshold percent
}
\description{
This function locates the minimum number of paramter/state variables that
must be eliminated base upon performance in order to decrease the domain
of the parameter by a specified percentage amount. This function links
performance of the y-values. Dependent on domainSize() function.
}
