% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfSurfacePreTwo.R
\name{discreteParameters}
\alias{discreteParameters}
\title{Performance Scaffold prerequisite functions}
\usage{
discreteParameters(par, divisions, min, max)
}
\arguments{
\item{par}{parameter state values (not performance/goodness of fit values)}

\item{divisions}{number of divisions into which the parameter space is discretised}

\item{min}{minimum parameter value in the domain}

\item{max}{maximum parameter value in the domain}
}
\value{
the identity of parameter state values in discrete bins
}
\description{
This function returns all of the parameter state values packaged into discrete intervals
for subsequently evaluating fit to generate the performance surface
}
