% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion}
\alias{dispersion}
\title{Dispersion}
\usage{
dispersion(sample, limDataframe)
}
\arguments{
\item{sample}{Resampled values from sirSampling}

\item{weighting}{Dilation parameter that increases the rate}
}
\value{
Parameter samples
}
\description{
Redistributing the directly resampled points within the discretised interval
}
