% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/domainSize.R
\name{domainSize}
\alias{domainSize}
\title{Size of the Domain of the Parameter Provided}
\usage{
domainSize(surface.x)
}
\arguments{
\item{surface.x}{x-values/state variables from the aggregate surface}
}
\value{
Size of Domain
}
\description{
Size of the domain of a parameter
}
