% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominator_constraint.R
\docType{data}
\name{dominator_constraint}
\alias{dominator_constraint}
\title{Restrict child position based on parent position}
\description{
This constraint requires children to be positioned at a certain side of their
parent and with a certain distance. It can be used to enforce a layering of
particles for e.g. DAG and tree layouts.
}
\section{Training parameters}{

The following parameters defines the training of the constraint and can be
passed along a call to \code{\link[=impose]{impose()}}
\itemize{
\item \code{distance} : The minimum orthogonal distance to the parent. Default to \code{0},
meaning that children are only required to be positioned to the specific
side of their parent. (\emph{tidy eval})
\item \code{angle} : The direction the children should be enforced to be relative to
their parent. Defaults to \code{-pi/2} which is equivalent to down. (\emph{tidy eval})
}
}

\seealso{
Other constraints: 
\code{\link{infinity_constraint}},
\code{\link{path_constraint}},
\code{\link{polygon_constraint}},
\code{\link{velocity_constraint}},
\code{\link{x_constraint}},
\code{\link{y_constraint}}
}
\concept{constraints}
\keyword{datasets}
