% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\title{Entropy of Dispersion}
\usage{
entropy(frequency)
}
\arguments{
\item{frequency}{Frequency vector (can be raw counts/goodness of fit values)}
}
\value{
Entropy/Concentration measure
}
\description{
This measurement of entropy estimates the dispersion of a vector - describing
the relative concentration of a vector's frequency into a number of age-classes
}
\examples{
entropy(c(1,0,0,0,0))
entropy(c(0.5, 0.5, 0.5, 0.5, 0.5))
entropy(c(0.95, 0.75, 0.55, 0.35, 0.15))
entropy(c(0.15, 0.35, 0.55, 0.75, 0.95))

}
