% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMestablishment.R
\name{establishment}
\alias{establishment}
\title{Establishment Probability}
\usage{
establishment(estabProbability)
}
\arguments{
\item{establishment}{probability of establishment/recruitment}
}
\value{
probability of establishment/recruitment
}
\description{
This function describes the probability than an individual born at time = t recruits to time t+1
}
