% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMfertility.R
\name{fertility}
\alias{fertility}
\title{Fertility Function}
\usage{
fertility(x0, fertInt, fertSlope)
}
\arguments{
\item{x0}{vector of individuals of state variable x (e.g., size) at time = t + 0 (new individuals at time t + 0)}

\item{fertInt}{intercept parameter of an exponential function}

\item{fertSlope}{slope parameter of an exponential function}
}
\value{
probability of reproducing for each individual of state variable x
}
\description{
This function describes the conditional size distribution of new offspring
(i.e., offspring born at time time = t when observed at time = t+1)
}
