% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggFormat.R
\name{ggFormat}
\alias{ggFormat}
\title{GG formatting}
\usage{
ggFormat(x0, x1, outerMatrix)
}
\arguments{
\item{x0}{state variable at time = t + 0 (vector)}

\item{x1}{state variable at time = t + 1 (vector)}

\item{outerMatrix}{matrix-formatted Kernel (U-, F-, or K-kernels)}
}
\value{
dataframe that can be passed to ggplot to be plotted as a heatmaap
}
\description{
This function transforms the matrix formatting to be plotted as a heatmap in ggplot
}
