% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infinity_constraint.R
\docType{data}
\name{infinity_constraint}
\alias{infinity_constraint}
\title{Reposition particles outside a canvas so they wrap around}
\description{
This constraint keeps particles inside of a defined area by positioning
exiting particles on the other side of the area. In effect this makes
particles that moves outside the upper bound reenter at the lower bound and
vice versa.
}
\section{Training parameters}{

The following parameters defines the training of the constraint and can be
passed along a call to \code{\link[=impose]{impose()}}
\itemize{
\item \code{xlim} : The left and right bound of the area
\item \code{ylim} : The upper and lower bound of the area
}
}

\seealso{
Other constraints: \code{\link{dominator_constraint}},
  \code{\link{path_constraint}},
  \code{\link{velocity_constraint}},
  \code{\link{x_constraint}}, \code{\link{y_constraint}}
}
\keyword{datasets}
