% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initiateSMC.R
\name{initiateSMC}
\alias{initiateSMC}
\title{Initiating Particle Distribution}
\usage{
initiateSMC(model.parameters, particle.number)
}
\arguments{
\item{parX}{parameter}
}
\value{
List containing
}
\description{
Defines the initial distribution of particles to be evaluated
}
