% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfSurfacePreOne.R
\name{interval}
\alias{interval}
\title{Performance Scaffold prerequisite functions}
\usage{
interval(par, divisions, position, min, max)
}
\arguments{
\item{par}{parameter state values (not performance/goodness of fit values)}

\item{divisions}{number of divisions into which the parameter space is discretised}

\item{position}{the rank-order integer of a parameter interval (see perfSurfacePreTwo)}

\item{min}{minimum parameter value in the domain}

\item{max}{maximum parameter value in the domain}
}
\value{
the identity of parameter state values that fall into the specified interval
}
\description{
This prerequisite function for the performance surface scaffolding
Associate sample/particles value with an individual parameter-value interval
}
