% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMboundary.R
\name{IPMboundary}
\alias{IPMboundary}
\title{Define the boundaries of the IPM}
\usage{
IPMboundary(
  minValue = NA,
  maxValue = NA,
  jointBuffer = NA,
  minBuffer = NA,
  maxBuffer = NA,
  minDomain = NA,
  maxDomain = NA
)
}
\arguments{
\item{minValue}{minimum value}

\item{maxValue}{maximum value}

\item{jointBuffer}{percentage buffer applied to minimum and maximum values - bypasses min/max buffer}

\item{minBuffer}{percentage buffer applied to the minimum value}

\item{minDomain}{lower floor on potential values - bypasses the values and buffer}

\item{maxDomain}{upper ceiling on potential values - bypasses the values and buffer}
}
\value{
Two dimensional boundary on the continuous variable of the integral population model
}
\description{
This function defines the limits of the Integral Population Model based on
The min/max values and optional buffer distance. Minimum and maximum values
can be based on empirical data using (min - or max(data, na.rm=TRUE)) or prescribed.
}
\examples{
IPMboundary(minValue = 2, maxValue = 8, jointBuffer = 0.2)
IPMboundary(minValue = 2, maxValue = 8, minBuffer = 0.2, maxBuffer = 0.05)
IPMboundary(minDomain = 1.5, maxDomain = 10)
}
