% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMdistance.R
\name{IPMdistance}
\alias{IPMdistance}
\title{IPM to Distance}
\usage{
IPMdistance(IPM.pre, observed.stage.dist, evaluate.params, i)
}
\arguments{
\item{IPM.pre}{IPM parameter dataframe with NA in parameter values that are unknowns}

\item{observed.stage.dist}{Dataframe of IPM parameters to be solved (cols = params, rows = samples/particles)}

\item{evaluate.params}{Parameters to evaluate in the parameter slots in the IPM (IPM.pre)}

\item{i}{sample/particle to be evaluated. This is the index that gets looped through with an apply function.}
}
\value{
Distance between the st/age distribution generated by the parameters and the reference distribution
}
\description{
This function takes IPM parameters, generates the IPM, generates the st/age distribution for each IPM,
and reports back the distance between the reference IPM and the true IPM. Requires that the stable-st/age
distribution be defined in the prior step
}
