% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMmesh.R
\name{IPMmesh}
\alias{IPMmesh}
\title{IPMmesh}
\usage{
IPMmesh(IPM.boundary, mesh.resoution)
}
\arguments{
\item{IPMboundary}{2-D vector or list of minimum and maximum values for the IPM}

\item{meshResoution}{Resolution of the IPM (number of grid-/mesh points across the domain)}
}
\value{
Vector of midpoints. n of the n x n matrix for which t -> t + 1 transitions will be evaluated.
}
\description{
This function defines the meshpoints of the IPM (grid resolution). Matrix of n x n at which
transition values are evaluated for each of the vital rate regression functions. Requires
IPM boundaries.
}
\examples{
IPMmesh(IPMboundary(minValue = 2, maxValue = 8, jointBuffer = 0.2), 100)
}
