% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMparameters.R
\name{IPMparameters}
\alias{IPMparameters}
\title{Parameter Reshaping for Kernel Generation}
\usage{
IPMparameters(param.dataframe)
}
\arguments{
\item{paramDataframe}{Dataframe parameter}
}
\value{
List of parameters
}
\description{
This function transforms the parameter dataframe into a list to be used
in the function to generate Kernels.
}
