% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterationStat.R
\name{iterationStat}
\alias{iterationStat}
\title{iterationStat}
\usage{
iterationStat(iteration, totalLength)
}
\arguments{
\item{iteration}{Looping variable (e.g., i in for(i in 1:100))}

\item{totalLength}{Total length of the iteration run}
}
\value{
Printed status update at a density of every fifth of the total progression.
}
\description{
This function updates the progress and estimated timeframe for the SMC
}
