% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadStatus.R
\name{loadStatus}
\alias{loadStatus}
\title{Check whether packages were successfully installed}
\usage{
loadStatus(x)
}
\arguments{
\item{loadingOutput}{output from sapply(packageNameVector, softloadpackages)}
}
\description{
Check if the packages were installed; flag any issues
}
\examples{
softloadpackages(dplyr)
}
