% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manybody_force.R
\docType{data}
\name{manybody_force}
\alias{manybody_force}
\title{Apply the mean velocity of all the neighbors to a particle}
\description{
This force takes the mean of all the neighbors (in the graph sense) of a
particle (and optionally itself) and applies it to itself.
}
\section{Training parameters}{

The following parameters defines the training of the force and can be passed
along a call to \code{\link[=wield]{wield()}}
\itemize{
\item \code{include_self} : Should the velocity of itself be included in the mean
calculation
\item \code{mode} : How should neighbors be found? \code{'all'} uses all edges. \code{'out'}
only uses outbound edges, and \code{'in'} only uses inbound edges. Ignored for
undirected particle graphs
}
}

\seealso{
Other forces: \code{\link{mean_force}},
  \code{\link{random_force}}
}
\keyword{datasets}
