% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mean_force.R
\docType{data}
\name{mean_force}
\alias{mean_force}
\title{Apply the mean velocity of all the neighbors to a particle}
\description{
This force takes the mean of all the neighbors (in the graph sense) of a
particle (and optionally itself) and applies it to itself.
}
\section{Training parameters}{

The following parameters defines the training of the force and can be passed
along a call to \code{\link[=wield]{wield()}}
\itemize{
\item \code{include_self} : Should the velocity of itself be included in the mean
calculation
\item \code{mode} : How should neighbors be found? \code{'all'} uses all edges. \code{'out'}
only uses outbound edges, and \code{'in'} only uses inbound edges. Ignored for
undirected particle graphs
}
}

\seealso{
Other forces: \code{\link{center_force}},
  \code{\link{collision_force}}, \code{\link{link_force}},
  \code{\link{manybody_force}}, \code{\link{random_force}},
  \code{\link{x_force}}, \code{\link{y_force}}
}
\keyword{datasets}
