% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelParams.R
\name{modelParams}
\alias{modelParams}
\title{Parameter File}
\usage{
modelParams(parX, ...)
}
\arguments{
\item{parX}{parameter list containing name and prior slot (output from solveParameter)}

\item{...}{additional parameter lists to be compiled}
}
\value{
List containing the names and priors for parameters to be solved
}
\description{
Identify the parameters to be solved for in the IPM
}
