% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfSurfaceMass.R
\name{multiSurface}
\alias{multiSurface}
\title{Aggregate Performance Surface}
\usage{
multiSurface(perf.dataframe, div)
}
\arguments{
\item{perf.dataframe}{dataframe containing the parameters and performance}

\item{div}{number of divisions into which performance is grouped/pooled}
}
\value{
List containing the x and y value surfaces for each parameter
}
\description{
This function takes teh
}
