% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramDecay.R
\name{paramDecay}
\alias{paramDecay}
\title{Parameter Decay Rate}
\usage{
paramDecay(surface.x, surface.y)
}
\arguments{
\item{surface.x}{Aggregate parameter state surface}

\item{surface.y}{Aggregate parameter state performance}
}
\value{
List of the decay of the domain. Each value represents the domain size
after eliminating n-number of elements from the surface.x parameter space based
on the rank performance
}
\description{
Generates the sequence at which the domain changes as one eliminates
parameter/state surface (surface.x) data based on performance (surface.y).
Dependent upon domainSize() function.
}
