% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_constraint.R
\docType{data}
\name{path_constraint}
\alias{path_constraint}
\title{Limit particle position to be along a path or outline}
\description{
This constraint repositions particles to their closest point along a given
path and sets their velocity to zero.
}
\section{Training parameters}{

The following parameters defines the training of the constraint and can be
passed along a call to \code{\link[=impose]{impose()}}
\itemize{
\item \code{path} : A two column matrix giving the path, or a
list of matrices to use multiple disconnected paths.
\item \code{closed} : Should the path close on itself. Defaults to \code{FALSE}
}
}

\seealso{
Other constraints: 
\code{\link{dominator_constraint}},
\code{\link{infinity_constraint}},
\code{\link{polygon_constraint}},
\code{\link{velocity_constraint}},
\code{\link{x_constraint}},
\code{\link{y_constraint}}
}
\concept{constraints}
\keyword{datasets}
