% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurfacePurturb.R
\name{perfPurturb}
\alias{perfPurturb}
\title{Performance Scaffold - Y-Values}
\usage{
perfPurturb(par, fit, divisions, min, max)
}
\arguments{
\item{par}{parameter value of the sample/particle}

\item{fit}{performance of the sample/particle}

\item{divisions}{number of divisions to bin into}

\item{min}{min parameter value to be evaluated}

\item{max}{max parameter value to be evaluated}
}
\value{
x-values at the mid-points of each bins for which performance is maximised
}
\description{
This function defines the performance measure evaluated over an interval defined by the
mid-point interval approximattion
}
