% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perfSurfaceX.R
\name{perfSurfaceX}
\alias{perfSurfaceX}
\title{Performance Scaffold}
\usage{
perfSurfaceX(min, max, divisions)
}
\arguments{
\item{min}{minimum domain value}

\item{max}{maximum domain value}

\item{divisions}{number of divisions to bin into}
}
\value{
x-values at the mid-points of each bins for which performance is maximised
}
\description{
This function defines the mid-point scaffolding across which the maximum performance
of a parameter value is assessed.
}
