% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polygon_constraint.R
\docType{data}
\name{polygon_constraint}
\alias{polygon_constraint}
\title{Fixes particles to be inside a polygon}
\description{
This constraint prevents particles from moving outside of one or more
polygons. If a particle ventures outside it will be moved back to its closest
point inside the specified polygon(s) and have its velocity set to zero.
}
\section{Training parameters}{

The following parameters defines the training of the constraint and can be
passed along a call to \code{\link[=impose]{impose()}}
\itemize{
\item \code{polygon} : A two column matrix giving the polygon, or a list of matrices
to use multiple polygons. Overlapping polygons will be subtracted from each
other so it is possible to define polygons with holes.
}
}

\seealso{
Other constraints: 
\code{\link{dominator_constraint}},
\code{\link{infinity_constraint}},
\code{\link{path_constraint}},
\code{\link{velocity_constraint}},
\code{\link{x_constraint}},
\code{\link{y_constraint}}
}
\concept{constraints}
\keyword{datasets}
