% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printNames.R
\name{printNames}
\alias{printNames}
\title{Printing parameter names}
\usage{
printNames(default = NULL)
}
\arguments{
\item{default}{NULL}
}
\value{
List of parameters
}
\description{
Prints the parameter names for quick reference
}
