% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_force.R
\docType{data}
\name{random_force}
\alias{random_force}
\title{Modify the velocity randomly at each step}
\description{
This force applies a random velocity modification to all particles. The
moodfication is uniformly distributed and bound be the parameters provided
during initialisation.
}
\section{Training parameters}{

The following parameters defines the training of the force and can be passed
along a call to \code{\link[=wield]{wield()}}
\itemize{
\item \code{xmin, xmax} : The bounds of the modification in the horizontal direction
\item \code{ymin, ymax} : The bounds of the modification in the vertical direction
}
}

\seealso{
Other forces: \code{\link{center_force}},
  \code{\link{collision_force}}, \code{\link{link_force}},
  \code{\link{manybody_force}}, \code{\link{mean_force}},
  \code{\link{reset_force}}, \code{\link{trap_force}},
  \code{\link{x_force}}, \code{\link{y_force}}
}
\keyword{datasets}
