% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/modify.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{filter}
\alias{slice}
\alias{mutate}
\alias{mutate_at}
\alias{mutate_all}
\alias{activate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}

  \item{tidygraph}{\code{\link[tidygraph]{activate}}, \code{\link[tidygraph:reexports]{filter}}, \code{\link[tidygraph:reexports]{mutate}}, \code{\link[tidygraph:reexports]{mutate_all}}, \code{\link[tidygraph:reexports]{mutate_at}}, \code{\link[tidygraph:reexports]{slice}}}
}}

