% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IPMreproduction.R
\name{reproduction}
\alias{reproduction}
\title{Reproductive Function}
\usage{
reproduction(
  x1,
  x0,
  estabProbability,
  recruitMean,
  recruitSD,
  fertInt,
  fertSlope
)
}
\arguments{
\item{x1}{vector of individuals of state variable x (e.g., size) at time = t + 1 (new individuals at time t + 1)}

\item{x0}{vector of individuals of state variable x (e.g., size) at time = t + 0 (new individuals at time t + 0)}

\item{recruitMean}{mean of the state variable (e.g., size) of new individuals recruited into the population at time t+1}

\item{recruitSD}{sd of the state variable (e.g., size) of new individuals recruited into the population at time t+1}

\item{fertInt}{intercept parameter of an exponential function}

\item{fertSlope}{slope parameter of an exponential function}

\item{establishment}{probability of establishment}
}
\value{
probability of reproducing for each individual of state variable x
}
\description{
This function evaluates (1) the state-specific fertility; (2) recruitment; (3) dispersion of new individuals.
}
