% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset_force.R
\docType{data}
\name{reset_force}
\alias{reset_force}
\title{Reset the velocity verlet of particles to a fixed value}
\description{
This force resets the velocity of particles at each generation. It can be
used if each generation should start from the same foundation rather than
accumulate as the simulation evolve. Particles where the parameters evaluates
to \code{NA} will ignore this force.
}
\section{Training parameters}{

The following parameters defines the training of the force and can be passed
along a call to \code{\link[=wield]{wield()}}
\itemize{
\item \code{xvel} : The x-velocity to reset to at each generation (\emph{tidy eval})
\item \code{yvel} : The y-velocity to reset to at each generation (\emph{tidy eval})
}
}

\seealso{
Other forces: \code{\link{center_force}},
  \code{\link{collision_force}}, \code{\link{field_force}},
  \code{\link{link_force}}, \code{\link{manybody_force}},
  \code{\link{map_force}}, \code{\link{mean_force}},
  \code{\link{random_force}}, \code{\link{trap_force}},
  \code{\link{x_force}}, \code{\link{y_force}}
}
\concept{forces}
\keyword{datasets}
