% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seriesA.R
\name{seriesA}
\alias{seriesA}
\title{A working dataset for parameterisation}
\usage{
seriesA(dataframe)
}
\arguments{
\item{dataframe}{blank dataframe}
}
\description{
A working dataset for parameterisation
}
\examples{
softloadpackages(dplyr)
}
