% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShannonDistance.R
\name{ShannonDistance}
\alias{ShannonDistance}
\title{Inverse Jensen-Shannon Divergence}
\usage{
ShannonDistance(x, y)
}
\arguments{
\item{x0}{reference st/age distribution (i.e., true st/age distribution)}

\item{x1}{observed st/age distribution}
}
\value{
Jensen Shannon Distance/goodness of fit measure
}
\description{
Jensen-Shannon Divergence (JSD) offers an unbiased, symmetric distance measure between two
statistical distribution. When distributions are identical JSD = 0. When distributions differ
JSD increases with no theoretical ceiling - similar to Shannon entropy.
}
