% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirSampling.R
\name{sirSampling}
\alias{sirSampling}
\title{Resampling function}
\usage{
sirSampling(resample.dataframe, resample.weight)
}
\arguments{
\item{resample.dataframe}{Dataframe of parameters on the restricted parameters}

\item{weighting}{Dilation parameter that increases the rate}
}
\value{
New series of samples
}
\description{
Sample Importance Resampling implementation function - Note the new dependency on
dispersion() function to redistribute the samples after the raw aggregate point
recounting is completed.
}
