% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/softloadpackages.R
\name{softloadPackages}
\alias{softloadPackages}
\title{Install Packages that are not already installed}
\usage{
softloadPackages(packagename)
}
\arguments{
\item{packageName}{name of package}
}
\description{
This function loads any packages that are not already installed in one's files
}
\examples{
softloadpackages(dplyr)
}
