% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solveParameter.R
\name{solveParameter}
\alias{solveParameter}
\title{Generate a list with the parameter name and prior}
\usage{
solveParameter(paramName)
}
\arguments{
\item{paramName}{IPM parameter to be solved}
}
\value{
List with the IPM parameter name attached to a prior
}
\description{
This function creates a list containing the parameter name and a placehold for a file
}
