% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stableStage.R
\name{stableStage}
\alias{stableStage}
\title{Stable Stage Distribution}
\usage{
stableStage(A)
}
\arguments{
\item{A}{Integral Population Model}
}
\value{
stable stage distribution
}
\description{
Calculates SSD from the
}
