% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subDomain.R
\name{subDomain}
\alias{subDomain}
\title{subDomain}
\usage{
subDomain(surface.object, cut.threshold)
}
\arguments{
\item{surface.object}{surface object from perfSurfaceMass}

\item{cut.threshold}{number of divisions into which performance is grouped/pooled}
}
\value{
List containing the retricted parameter values and corresponding fit values
}
\description{
This function uses the threshold and surface object to limit the domain.
}
