% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trueParameter.R
\name{trueParameter}
\alias{trueParameter}
\title{true Parameter}
\usage{
trueParameter(param.name, param.dataframe)
}
\arguments{
\item{param.name}{name(s) of the parameter(s) to be removed (atomic or vector )}

\item{param.dataframe}{dataframe containing the full set of parameters}
}
\value{
dataframe with the name and value of the values that have been knocked out of the IPM
}
\description{
This function extracts the parameter values for the parameter names
}
\examples{

}
