% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/velocity_constraint.R
\docType{data}
\name{velocity_constraint}
\alias{velocity_constraint}
\title{Fixes particles to a specific velocity range}
\description{
This constraint puts bounds on the magnitude of velocity a particle can have.
Particles where either end of the bound is NA ignores the constraint. If a
particle with no velocity is forced to have a velocity the direction will be
random.
}
\section{Training parameters}{

The following parameters defines the training of the constraint and can be
passed along a call to \code{\link[=impose]{impose()}}
\itemize{
\item \code{v} : The velocity allowed for the particle.
\item \code{vmin} : The lowest permissable velocity. If \code{NULL} then \code{v} will be used.
\item \code{vmax} : The highest permissable velocity. If \code{NULL} then \code{v} will be used.
}
}

\seealso{
Other constraints: \code{\link{dominator_constraint}},
  \code{\link{infinity_constraint}},
  \code{\link{path_constraint}},
  \code{\link{polygon_constraint}},
  \code{\link{x_constraint}}, \code{\link{y_constraint}}
}
\concept{constraints}
\keyword{datasets}
