% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wield.R
\name{wield}
\alias{wield}
\title{Assign a force to a simulation}
\usage{
wield(simulation, force, ...)
}
\arguments{
\item{simulation}{A simulation object}

\item{force}{A force object}

\item{...}{Parameters passed on to the training of the force}
}
\value{
A simulation with the force added
}
\description{
This function adds a new force to the simulation and trains the force on the
current particle graph. The parameters passed on to the force training are
using tidy evaluation from the rlang package. Depending on the force the
data getting referenced is either the node or the edge data of the particle
graph.
}
\examples{
graph <- igraph::make_graph('folkman')
graph \%>\%
  simulate() \%>\%
  wield(link_force)

}
