% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_constraint.R
\docType{data}
\name{x_constraint}
\alias{x_constraint}
\title{Fixes particles to a horizontal position}
\description{
This constraint simply prevents particles from moving in the x direction. For
particles where the constraint evaluates to \code{NA} this constraint is ignored.
If the constraint is enforced the velocity in the x direction will be set to
\code{0}.
}
\section{Training parameters}{

The following parameters defines the training of the constraint and can be
passed along a call to \code{\link[=impose]{impose()}}
\itemize{
\item \code{x} : The position on the x-axis to fix to.
\item \code{xmin} : The lowest permissable x-value. If \code{NULL} then \code{x} will be used.
\item \code{xmax} : The highest permissable x-value. If \code{NULL} then \code{x} will be used.
}
}

\seealso{
Other constraints: \code{\link{dominator_constraint}},
  \code{\link{infinity_constraint}},
  \code{\link{path_constraint}},
  \code{\link{polygon_constraint}},
  \code{\link{velocity_constraint}},
  \code{\link{y_constraint}}
}
\concept{constraints}
\keyword{datasets}
