% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_force.R
\docType{data}
\name{x_force}
\alias{x_force}
\title{Attract particles towards a horizontal position}
\description{
This force simply pulls particles towards a fixed position on the x-axis.
}
\section{Training parameters}{

The following parameters defines the training of the force and can be passed
along a call to \code{\link[=wield]{wield()}}
\itemize{
\item \code{strength} : The strength with which the attraction occurs (\emph{tidy eval})
\item \code{x} : The position on the x-axis to pull towards. (\emph{tidy eval})
}
}

\seealso{
Other forces: 
\code{\link{center_force}},
\code{\link{collision_force}},
\code{\link{field_force}},
\code{\link{link_force}},
\code{\link{manybody_force}},
\code{\link{map_force}},
\code{\link{mean_force}},
\code{\link{random_force}},
\code{\link{reset_force}},
\code{\link{trap_force}},
\code{\link{y_force}}
}
\concept{forces}
\keyword{datasets}
