% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/particles-package.R
\docType{package}
\name{particles-package}
\title{particles: A graph based particle simulator based on d3-force}
\description{
\if{html}{\figure{logo.png}{options: align='right'}}

Simulating particle movement in 2D space has many application. The
'particles' package implements a particle simulator based on the ideas
behind the 'd3-force' javascript library. 'particles' implements all forces
defined in 'd3-force' as well as others such as vector fields, traps, and
attractors.
}
