% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modify.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{filter}
\alias{slice}
\alias{mutate}
\alias{mutate_at}
\alias{mutate_all}
\alias{activate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{mutate_all}}, \code{\link[dplyr:mutate_all]{mutate_at}}, \code{\link[dplyr]{slice}}}

  \item{tidygraph}{\code{\link[tidygraph]{activate}}}
}}

